refactor(runtime): extract subagent runner to orchestrator/subagent module
Move the SubagentRunner trait implementation (run_subagent, cancel_agent)
from orchestrator/mod.rs into a dedicated orchestrator/subagent.rs file.
This reduces mod.rs from 2,276 to 1,800 lines while keeping the subagent
logic self-contained and easier to test independently.

Clean up now-unused imports in mod.rs (AgentReport, AgentReportStatus,
AgentSpawn, SubagentRunner, DEFAULT_MAX_AGENT_DEPTH, async_trait).
